Restructured adjusters
