+++ app/lib/business/operations/task_adjusters/dual_simplex_adjuster.dart
@@ -1,7 +1,7 @@
 import 'composite_adjuster.dart';
 import 'extremum_to_minimum_adjuster.dart';
-import 'dual_simplex_restrictions_to_equalities_adjuster.dart';
-import 'dual_simplex_basis_creator.dart';
+import 'restrictions/dual_simplex_restrictions_to_equalities_adjuster.dart';
+import 'restrictions/dual_simplex_basis_creator.dart';
 
 class DualSimplexAdjuster extends CompositeAdjuster {
   const DualSimplexAdjuster()
