+++ app/lib/business/operations/task_adjusters/restrictions/dual_simplex_restrictions_to_equalities_adjuster.dart
