+++ app/lib/business/operations/task_adjusters/simplex_adjuster.dart
@@ -1,16 +1,16 @@
 import 'package:app/business/operations/task_adjusters/extremum_to_minimum_adjuster.dart';
+import 'composite_adjuster.dart';
 import 'restrictions/equal_restrictions_adjuster.dart';
 import 'restrictions/greater_equal_restrictions_adjuster.dart';
 import 'restrictions/lower_equal_restrictions_adjuster.dart';
-import 'simplex_restrictions_free_members_adjuster.dart';
-import 'composite_adjuster.dart';
+import 'restrictions/positive_free_member_adjuster.dart';
 
 class SimplexAdjuster extends CompositeAdjuster {
   const SimplexAdjuster()
       : super(
           const [
             ExtremumToMinimumAdjuster(),
-            SimplexRestrictionsFreeMemberAdjuster(),
+            PositiveFreeMemberAdjuster(),
             GreaterEqualRestrictionsAdjuster(),
             EqualRestrictionsAdjuster(),
             LowerEqualRestrictionsAdjuster(),
